# enterprise/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Boundless Enterprise Multipass API
  description: |
    Enterprise-grade unified access and control layer for the Boundless blockchain platform.

    ## Features
    - Post-Quantum Cryptography (Dilithium5 + Kyber1024)
    - KYC/AML Identity Management
    - Multi-Asset Wallet System
    - Internal Asset Trading Market
    - Hardware Pass NFC Integration
    - SSO & Session Management
    - Enterprise Application Registry
    - Events & Reporting

    ## Security
    - All endpoints (except `/api/auth/login` and `/api/auth/register`) require authentication
    - JWT-based session tokens
    - Rate limiting on sensitive endpoints
    - Encrypted private key storage with master encryption key

    ## Base URL
    `http://localhost:8080` (default development)

  version: 1.0.0
  contact:
    name: Boundless Platform
    url: https://boundless.example.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.boundless.example.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from `/api/auth/login`
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...
    AssetDefinition:
      type: object
      properties:
        asset_id:
          type: string
          format: uuid
        name:
          type: string
        symbol:
          type: string
        asset_type:
          type: string
          enum: [Native, UtilityToken, EquityToken, CarbonCredit, NFT, SubscriptionPass]
        total_supply:
          type: integer
          format: int64
        circulating_supply:
          type: integer
          format: int64
...
    Session:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        identity_id:
          type: string
          format: uuid
        token:
          type: string
          description: JWT authentication token
        expires_at:
          type: string
          format: date-time
paths:
...
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login user
      description: Authenticate with email and password, returns JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/assets/define:
    post:
      tags: [Assets]
      summary: Define new asset
      description: Create enterprise asset definition (tokens, carbon credits, NFTs, etc.)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, symbol, asset_type, total_supply]
              properties:
                name:
                  type: string
                symbol:
                  type: string
                asset_type:
                  type: string
                  enum: [Native, UtilityToken, EquityToken, CarbonCredit, NFT, SubscriptionPass]
                total_supply:
                  type: integer
                  format: int64
      responses:
        '200':
          description: Asset defined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetDefinition'
  /api/assets/list:
    get:
      tags: [Assets]
      summary: List all assets
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Asset list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssetDefinition'
  /api/assets/{asset_id}/issue:
    post:
      tags: [Assets]
      summary: Issue asset tokens
      description: Mint new tokens to a wallet
      security:
        - BearerAuth: []
      parameters:
        - name: asset_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wallet_id, quantity]
              properties:
                wallet_id:
                  type: string
                  format: uuid
                quantity:
                  type: integer
                  format: int64
      responses:
        '200':
          description: Asset issued
  /api/assets/{asset_id}/transfer:
    post:
      tags: [Assets]
      summary: Transfer asset
      description: Transfer asset tokens between wallets with blockchain settlement
      security:
        - BearerAuth: []
      parameters:
        - name: asset_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from_wallet_id, to_wallet_id, quantity]
              properties:
                from_wallet_id:
                  type: string
                  format: uuid
                to_wallet_id:
                  type: string
                  format: uuid
                quantity:
                  type: integer
                  format: int64
      responses:
        '200':
          description: Transfer successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  tx_hash:
                    type: string
                    description: Blockchain transaction hash

  # =============================================================================
  # MARKET
  # =============================================================================
...
security:
  - BearerAuth: []